Moved common visualelement handling functions to new file
